requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
docker==6.1.3
frozenlist==1.4.0
grpcio==1.56.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
pyvespa==0.34.0
requests==2.31.0
six==1.16.0
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.1
yarl==1.9.2
(venv) marcosesteve@mbp-de-marcos indoor-tiktok % pip freeze
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
docker==6.1.3
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
pyvespa==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.1
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
